% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{interval_width}
\alias{interval_width}
\title{Prediction Interval Widths}
\usage{
interval_width(object)
}
\arguments{
\item{object}{A \code{predictset_reg} object.}
}
\value{
A numeric vector of interval widths.
}
\description{
Returns the width of each prediction interval.
}
\examples{
set.seed(42)
x <- matrix(rnorm(200 * 3), ncol = 3)
y <- x[, 1] * 2 + rnorm(200)
x_new <- matrix(rnorm(50 * 3), ncol = 3)

result <- conformal_split(x, y, model = y ~ ., x_new = x_new)
widths <- interval_width(result)
summary(widths)

}
