% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.predictset_class}
\alias{print.predictset_class}
\title{Print Method for Classification Conformal Objects}
\usage{
\method{print}{predictset_class}(x, ...)
}
\arguments{
\item{x}{A \code{predictset_class} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for Classification Conformal Objects
}
