% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.predictset_model}
\alias{print.predictset_model}
\title{Print Method for Model Specifications}
\usage{
\method{print}{predictset_model}(x, ...)
}
\arguments{
\item{x}{A \code{predictset_model} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for Model Specifications
}
