% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.predictset_reg}
\alias{print.predictset_reg}
\title{Print Method for Regression Conformal Objects}
\usage{
\method{print}{predictset_reg}(x, ...)
}
\arguments{
\item{x}{A \code{predictset_reg} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for Regression Conformal Objects
}
