% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{set_size}
\alias{set_size}
\title{Prediction Set Sizes}
\usage{
set_size(object)
}
\arguments{
\item{object}{A \code{predictset_class} object.}
}
\value{
An integer vector of set sizes.
}
\description{
Returns the number of classes in each prediction set.
}
\examples{
set.seed(42)
n <- 300
x <- matrix(rnorm(n * 4), ncol = 4)
y <- factor(ifelse(x[,1] > 0, "A", "B"))
x_new <- matrix(rnorm(50 * 4), ncol = 4)

clf <- make_model(
  train_fun = function(x, y) glm(y ~ ., data = data.frame(y = y, x),
                                  family = "binomial"),
  predict_fun = function(object, x_new) {
    df <- as.data.frame(x_new)
    names(df) <- paste0("X", seq_len(ncol(x_new)))
    p <- predict(object, newdata = df, type = "response")
    cbind(A = 1 - p, B = p)
  },
  type = "classification"
)

result <- conformal_lac(x, y, model = clf, x_new = x_new)
sizes <- set_size(result)
table(sizes)

}
