% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.predictset_class}
\alias{summary.predictset_class}
\title{Summary Method for Classification Conformal Objects}
\usage{
\method{summary}{predictset_class}(object, ...)
}
\arguments{
\item{object}{A \code{predictset_class} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Summary Method for Classification Conformal Objects
}
