% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.predictset_reg}
\alias{summary.predictset_reg}
\title{Summary Method for Regression Conformal Objects}
\usage{
\method{summary}{predictset_reg}(object, ...)
}
\arguments{
\item{object}{A \code{predictset_reg} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Summary Method for Regression Conformal Objects
}
